Currently displays weather JSON data.
First what I would like to do is test to see if JSON can be called correctly.

Second, I would like to add a way for the prompt to enter the users address through an inbound text message to the server.

Finally, I would like that data to be cached so the user does not have to enter the data again and it registers the user. Maybe create a database of 'Registered Users' pointing to a Key Value Pair of Phone Number to Area Code/City. How will that scale to determine what they're growing? I have no idea yet. Let's take it a step at a time.